+++ src/test/java/org/realityforge/guiceyloops/server/TypeListenerTest.java
@@ -4,12 +4,32 @@ import com.google.inject.AbstractModule;
 import com.google.inject.Guice;
 import com.google.inject.Injector;
 import com.google.inject.name.Names;
+import java.io.Serializable;
 import java.lang.reflect.Field;
 import java.util.List;
 import java.util.Map;
 import javax.annotation.Nullable;
 import javax.annotation.Resource;
 import javax.ejb.EJB;
+import javax.inject.Inject;
+import javax.jms.BytesMessage;
+import javax.jms.ConnectionMetaData;
+import javax.jms.Destination;
+import javax.jms.ExceptionListener;
+import javax.jms.JMSConnectionFactory;
+import javax.jms.JMSConsumer;
+import javax.jms.JMSContext;
+import javax.jms.JMSProducer;
+import javax.jms.MapMessage;
+import javax.jms.Message;
+import javax.jms.ObjectMessage;
+import javax.jms.Queue;
+import javax.jms.QueueBrowser;
+import javax.jms.StreamMessage;
+import javax.jms.TemporaryQueue;
+import javax.jms.TemporaryTopic;
+import javax.jms.TextMessage;
+import javax.jms.Topic;
 import javax.persistence.EntityGraph;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
@@ -48,6 +68,7 @@ public final class TypeListenerTest
                          "_entityManagerBViaPersistenceContextWithKey",
                          TestEntityManager2.class );
     assertFieldInjected( instance, ComponentA.class, "_webServiceType" );
+    assertFieldInjected( instance, ComponentA.class, "_context", TestJMSContext.class );
   }
 
   @Test
@@ -67,6 +88,7 @@ public final class TypeListenerTest
                          "_entityManagerBViaPersistenceContextWithKey",
                          TestEntityManager2.class );
     assertFieldInjected( instance, ComponentA.class, "_webServiceType" );
+    assertFieldInjected( instance, ComponentA.class, "_context", TestJMSContext.class );
   }
 
   private void assertFieldNotInjected( final Object instance, final Class declaringType, final String fieldName )
@@ -118,6 +140,8 @@ public final class TypeListenerTest
   {
     protected void configure()
     {
+      bind( JMSContext.class ).to( TestJMSContext2.class );
+      bind( JMSContext.class ).annotatedWith( Names.named( "jms/Blah" ) ).to( TestJMSContext.class );
       bind( EntityManager.class ).to( TestEntityManager.class );
       bind( EntityManager.class ).annotatedWith( Names.named( "X" ) ).toInstance( new TestEntityManager2() );
       bind( ComponentB.class ).annotatedWith( Names.named( "some/resource/name" ) ).toInstance( new ComponentB() );
@@ -140,6 +164,10 @@ public final class TypeListenerTest
 
     @Resource(lookup = "some/resource/name")
     private ComponentB _componentBViaResourceLookup;
+
+    @Inject
+    @JMSConnectionFactory( "jms/Blah" )
+    private JMSContext _context;
 
     //Should not be injected
     @PersistenceContext
@@ -439,6 +467,510 @@ public final class TypeListenerTest
     public Metamodel getMetamodel()
     {
       return null;
+    }
+  }
+
+  static final class TestJMSContext
+    implements JMSContext
+  {
+    @Override
+    public JMSContext createContext( final int sessionMode )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSProducer createProducer()
+    {
+      return null;
+    }
+
+    @Override
+    public String getClientID()
+    {
+      return null;
+    }
+
+    @Override
+    public void setClientID( final String clientID )
+    {
+    }
+
+    @Override
+    public ConnectionMetaData getMetaData()
+    {
+      return null;
+    }
+
+    @Override
+    public ExceptionListener getExceptionListener()
+    {
+      return null;
+    }
+
+    @Override
+    public void setExceptionListener( final ExceptionListener listener )
+    {
+    }
+
+    @Override
+    public void start()
+    {
+    }
+
+    @Override
+    public void stop()
+    {
+    }
+
+    @Override
+    public void setAutoStart( final boolean autoStart )
+    {
+    }
+
+    @Override
+    public boolean getAutoStart()
+    {
+      return false;
+    }
+
+    @Override
+    public void close()
+    {
+    }
+
+    @Override
+    public BytesMessage createBytesMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public MapMessage createMapMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public Message createMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public ObjectMessage createObjectMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public ObjectMessage createObjectMessage( final Serializable object )
+    {
+      return null;
+    }
+
+    @Override
+    public StreamMessage createStreamMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public TextMessage createTextMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public TextMessage createTextMessage( final String text )
+    {
+      return null;
+    }
+
+    @Override
+    public boolean getTransacted()
+    {
+      return false;
+    }
+
+    @Override
+    public int getSessionMode()
+    {
+      return 0;
+    }
+
+    @Override
+    public void commit()
+    {
+    }
+
+    @Override
+    public void rollback()
+    {
+    }
+
+    @Override
+    public void recover()
+    {
+    }
+
+    @Override
+    public JMSConsumer createConsumer( final Destination destination )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createConsumer( final Destination destination, final String messageSelector )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createConsumer( final Destination destination,
+                                       final String messageSelector,
+                                       final boolean noLocal )
+    {
+      return null;
+    }
+
+    @Override
+    public Queue createQueue( final String queueName )
+    {
+      return null;
+    }
+
+    @Override
+    public Topic createTopic( final String topicName )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createDurableConsumer( final Topic topic, final String name )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createDurableConsumer( final Topic topic,
+                                              final String name,
+                                              final String messageSelector,
+                                              final boolean noLocal )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createSharedDurableConsumer( final Topic topic, final String name )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createSharedDurableConsumer( final Topic topic, final String name, final String messageSelector )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createSharedConsumer( final Topic topic, final String sharedSubscriptionName )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createSharedConsumer( final Topic topic,
+                                             final String sharedSubscriptionName,
+                                             final String messageSelector )
+    {
+      return null;
+    }
+
+    @Override
+    public QueueBrowser createBrowser( final Queue queue )
+    {
+      return null;
+    }
+
+    @Override
+    public QueueBrowser createBrowser( final Queue queue, final String messageSelector )
+    {
+      return null;
+    }
+
+    @Override
+    public TemporaryQueue createTemporaryQueue()
+    {
+      return null;
+    }
+
+    @Override
+    public TemporaryTopic createTemporaryTopic()
+    {
+      return null;
+    }
+
+    @Override
+    public void unsubscribe( final String name )
+    {
+    }
+
+    @Override
+    public void acknowledge()
+    {
+    }
+  }
+
+  static final class TestJMSContext2
+    implements JMSContext
+  {
+    @Override
+    public JMSContext createContext( final int sessionMode )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSProducer createProducer()
+    {
+      return null;
+    }
+
+    @Override
+    public String getClientID()
+    {
+      return null;
+    }
+
+    @Override
+    public void setClientID( final String clientID )
+    {
+    }
+
+    @Override
+    public ConnectionMetaData getMetaData()
+    {
+      return null;
+    }
+
+    @Override
+    public ExceptionListener getExceptionListener()
+    {
+      return null;
+    }
+
+    @Override
+    public void setExceptionListener( final ExceptionListener listener )
+    {
+    }
+
+    @Override
+    public void start()
+    {
+    }
+
+    @Override
+    public void stop()
+    {
+    }
+
+    @Override
+    public void setAutoStart( final boolean autoStart )
+    {
+    }
+
+    @Override
+    public boolean getAutoStart()
+    {
+      return false;
+    }
+
+    @Override
+    public void close()
+    {
+    }
+
+    @Override
+    public BytesMessage createBytesMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public MapMessage createMapMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public Message createMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public ObjectMessage createObjectMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public ObjectMessage createObjectMessage( final Serializable object )
+    {
+      return null;
+    }
+
+    @Override
+    public StreamMessage createStreamMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public TextMessage createTextMessage()
+    {
+      return null;
+    }
+
+    @Override
+    public TextMessage createTextMessage( final String text )
+    {
+      return null;
+    }
+
+    @Override
+    public boolean getTransacted()
+    {
+      return false;
+    }
+
+    @Override
+    public int getSessionMode()
+    {
+      return 0;
+    }
+
+    @Override
+    public void commit()
+    {
+    }
+
+    @Override
+    public void rollback()
+    {
+    }
+
+    @Override
+    public void recover()
+    {
+    }
+
+    @Override
+    public JMSConsumer createConsumer( final Destination destination )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createConsumer( final Destination destination, final String messageSelector )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createConsumer( final Destination destination,
+                                       final String messageSelector,
+                                       final boolean noLocal )
+    {
+      return null;
+    }
+
+    @Override
+    public Queue createQueue( final String queueName )
+    {
+      return null;
+    }
+
+    @Override
+    public Topic createTopic( final String topicName )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createDurableConsumer( final Topic topic, final String name )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createDurableConsumer( final Topic topic,
+                                              final String name,
+                                              final String messageSelector,
+                                              final boolean noLocal )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createSharedDurableConsumer( final Topic topic, final String name )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createSharedDurableConsumer( final Topic topic, final String name, final String messageSelector )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createSharedConsumer( final Topic topic, final String sharedSubscriptionName )
+    {
+      return null;
+    }
+
+    @Override
+    public JMSConsumer createSharedConsumer( final Topic topic,
+                                             final String sharedSubscriptionName,
+                                             final String messageSelector )
+    {
+      return null;
+    }
+
+    @Override
+    public QueueBrowser createBrowser( final Queue queue )
+    {
+      return null;
+    }
+
+    @Override
+    public QueueBrowser createBrowser( final Queue queue, final String messageSelector )
+    {
+      return null;
+    }
+
+    @Override
+    public TemporaryQueue createTemporaryQueue()
+    {
+      return null;
+    }
+
+    @Override
+    public TemporaryTopic createTemporaryTopic()
+    {
+      return null;
+    }
+
+    @Override
+    public void unsubscribe( final String name )
+    {
+    }
+
+    @Override
+    public void acknowledge()
+    {
     }
   }
 }